Fix issue #7372, move widget functions into matplotlib.testing.widgets.
